Tidied up model + form fields and validators
Also moved `MultiLingualFieldAdmin` from `multilingual/modelfields.py` to `multilingual/admin.py`.
