+++ make_queue/admin.py
@@ -7,7 +7,7 @@ from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 
 from users.admin import get_user_search_fields
-from web.multilingual.database import MultiLingualFieldAdmin
+from web.multilingual.admin import MultiLingualFieldAdmin
 from .models.course import Printer3DCourse
 from .models.models import Machine, MachineType, MachineUsageRule, Quota, Reservation, ReservationQuerySet, ReservationRule
