+++ news/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from web.multilingual.database import MultiLingualFieldAdmin
+from web.multilingual.admin import MultiLingualFieldAdmin
 from .models import Article, Event, EventTicket, TimePlace
 
 admin.site.register(Article, MultiLingualFieldAdmin)
