+++ users/models.py
@@ -2,7 +2,7 @@ from django.contrib.auth.models import AbstractUser
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from card.models import CardNumberField
+from card.modelfields import CardNumberField
 
 
 class User(AbstractUser):
